refactor: move new BaseTypes into overlays/shared
